Consolidate service registry
