+++ apps/openchallenges/config-server/project.json
@@ -77,8 +77,6 @@
   },
   "tags": ["type:service", "scope:backend"],
   "implicitDependencies": [
-    "openchallenges-app-config-data",
-    "openchallenges-service-registry",
-    "openchallenges-vault"
+    "openchallenges-app-config-data"
   ]
 }
