+++ apps/openchallenges/service-registry/.gitignore
@@ -2,6 +2,7 @@ HELP.md
 target/
 !**/src/main/**/target/
 !**/src/test/**/target/
+log/
 
 ### STS ###
 .apt_generated
