+++ apps/openchallenges/service-registry/gradle.properties
@@ -1,3 +1,4 @@
+lombokVersion=1.18.24
 openchallengesVersion=0.0.1-SNAPSHOT
 springBootVersion=2.7.8
 springCloudVersion=3.1.4
