--- apps/openchallenges/service-registry/src/main/java/org/sagebionetworks/challenge/ChallengeServiceRegistryApplication.java
@@ -1,15 +1,0 @@
-package org.sagebionetworks.challenge;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@EnableEurekaServer
-@SpringBootApplication
-public class ChallengeServiceRegistryApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(ChallengeServiceRegistryApplication.class, args);
-  }
-
-}
