+++ libs/openchallenges/app-config-data/src/main/java/org/sagebionetworks/openchallenges/app/config/data/ServiceRegistryConfigData.java
@@ -1,0 +1,13 @@
+package org.sagebionetworks.openchallenges.app.config.data;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Data
+@Configuration
+@ConfigurationProperties(prefix = "openchallenges-service-registry")
+public class ServiceRegistryConfigData {
+
+  private String welcomeMessage;
+}
